refactor proxy module boundaries
