+++ crates/token_proxy_core/src/proxy/codex_compat/mod.rs
@@ -40,5 +40,4 @@ pub(crate) fn apply_codex_headers_if_needed(
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "codex_compat.test.rs"]
 mod tests;
