+++ crates/token_proxy_core/src/proxy/codex_compat/tests.rs
