+++ crates/token_proxy_core/src/proxy/openai_compat/mod.rs
@@ -961,5 +961,4 @@ fn responses_output_text_part(text: String, annotations: Vec<Value>) -> Value {
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "openai_compat.test.rs"]
 mod tests;
