+++ crates/token_proxy_core/src/proxy/request_token_estimate/mod.rs
@@ -552,5 +552,4 @@ enum MediaKind {
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "request_token_estimate.test.rs"]
 mod tests;
