+++ crates/token_proxy_core/src/proxy/response/mod.rs
@@ -292,5 +292,4 @@ pub(crate) use streaming::STREAM_DROPPED_ERROR;
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "response.test.rs"]
 mod tests;
