+++ crates/token_proxy_core/src/proxy/server/mod.rs
@@ -160,5 +160,4 @@ async fn proxy_request(
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "server.test.rs"]
 mod tests;
