+++ crates/token_proxy_core/src/proxy/server_helpers/mod.rs
@@ -414,5 +414,4 @@ fn ensure_json_object_field<'a>(
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "server_helpers.test.rs"]
 mod tests;
