+++ crates/token_proxy_core/src/proxy/usage/mod.rs
@@ -186,5 +186,4 @@ fn update_usage(snapshot: &mut UsageSnapshot, data: &str) {
 
 // 单元测试拆到独立文件，使用 `#[path]` 以保持 `.test.rs` 命名约定。
 #[cfg(test)]
-#[path = "usage.test.rs"]
 mod tests;
